requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
fastapi==0.115.4
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.13
PySocks==1.7.1
pytest==8.3.3
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.26.1
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.41.2
swarmauri==0.5.0
swarmauri-core==0.5.0
threadpoolctl==3.5.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.0
wsproto==1.2.0
